Move base64 classes into a base64 folder
